isolate inmemory storage implementations
