+++ src/eth/storage/inmemory/inmemory_permanent.rs
@@ -11,15 +11,13 @@ use tokio::sync::RwLock;
 use tokio::sync::RwLockReadGuard;
 use tokio::sync::RwLockWriteGuard;
 
-use super::inmemory_account::InMemoryAccount;
+use super::inmemory_account::InMemoryAccountPermanent;
+use super::inmemory_state::StorageState;
 use crate::eth::primitives::Account;
 use crate::eth::primitives::Address;
 use crate::eth::primitives::Block;
 use crate::eth::primitives::BlockNumber;
 use crate::eth::primitives::BlockSelection;
-use crate::eth::primitives::Execution;
-use crate::eth::primitives::ExecutionConflicts;
-use crate::eth::primitives::ExecutionConflictsBuilder;
 use crate::eth::primitives::Hash;
 use crate::eth::primitives::LogFilter;
 use crate::eth::primitives::LogMined;
@@ -27,17 +25,30 @@ use crate::eth::primitives::Slot;
 use crate::eth::primitives::SlotIndex;
 use crate::eth::primitives::StoragePointInTime;
 use crate::eth::primitives::TransactionMined;
-use crate::eth::storage::inmemory::InMemoryAccountPermanent;
-use crate::eth::storage::inmemory::InMemoryAccountTemporary;
 use crate::eth::storage::PermanentStorage;
 use crate::eth::storage::StorageError;
-use crate::eth::storage::TemporaryStorage;
 
-/// The inmemory storage is split into two structs. InMemoryStoragePermanent and
-/// InMemoryStorageTemporary to facilitate debugging when using the inmemory storage
-/// for both temp and perm contexts.
 
-/// In-memory implementation using maps.
+#[derive(Debug, Default)]
+struct InMemoryPermanentStorageState {
+    accounts: HashMap<Address, InMemoryAccountPermanent>,
+    transactions: HashMap<Hash, TransactionMined>,
+    blocks_by_number: IndexMap<BlockNumber, Arc<Block>>,
+    blocks_by_hash: IndexMap<Hash, Arc<Block>>,
+    logs: Vec<LogMined>,
+}
+
+
+impl StorageState<InMemoryAccountPermanent> for InMemoryPermanentStorageState {
+    fn get_accounts(&self) -> &HashMap<Address, InMemoryAccountPermanent> {
+        &self.accounts
+    }
+
+    fn get_accounts_mut(&mut self) -> &mut HashMap<Address, InMemoryAccountPermanent> {
+        &mut self.accounts
+    }
+}
+
 #[derive(Debug)]
 pub struct InMemoryStoragePermanent {
     state: RwLock<InMemoryPermanentStorageState>,
@@ -66,96 +77,7 @@ impl InMemoryStoragePermanent {
     }
 }
 
-trait StorageState<T: InMemoryAccount> {
-    fn get_accounts(&self) -> &HashMap<Address, T>;
-    fn get_accounts_mut(&mut self) -> &mut HashMap<Address, T>;
 
-    fn save_account_changes(&mut self, block_number: BlockNumber, execution: Execution) {
-        let is_success = execution.is_success();
-        for changes in execution.changes {
-            let account = self
-                .get_accounts_mut()
-                .entry(changes.address.clone())
-                .or_insert_with(|| T::new(changes.address));
-
-            // account basic info
-            if let Some(nonce) = changes.nonce.take() {
-                account.set_nonce(block_number, nonce);
-            }
-            if let Some(balance) = changes.balance.take() {
-                account.set_balance(block_number, balance);
-            }
-            if let Some(Some(bytecode)) = changes.bytecode.take() {
-                account.set_bytecode(block_number, bytecode);
-            }
-
-            // slots
-            if is_success {
-                for (_, slot) in changes.slots {
-                    if let Some(slot) = slot.take_modified() {
-                        account.set_slot(block_number, slot);
-                    }
-                }
-            }
-        }
-    }
-
-    fn check_conflicts(&self, execution: &Execution) -> Option<ExecutionConflicts> {
-        let mut conflicts = ExecutionConflictsBuilder::default();
-
-        for change in &execution.changes {
-            let address = &change.address;
-
-            if let Some(account) = self.get_accounts().get(address) {
-                // check account info conflicts
-                if let Some(touched_nonce) = change.nonce.take_original_ref() {
-                    let nonce = account.get_current_nonce();
-                    if touched_nonce != nonce {
-                        conflicts.add_nonce(address.clone(), nonce.clone(), touched_nonce.clone());
-                    }
-                }
-                if let Some(touched_balance) = change.balance.take_original_ref() {
-                    let balance = account.get_current_balance();
-                    if touched_balance != balance {
-                        conflicts.add_balance(address.clone(), balance.clone(), touched_balance.clone());
-                    }
-                }
-
-                // check slots conflicts
-                for (touched_slot_index, touched_slot) in &change.slots {
-                    if let Some(slot) = account.get_current_slot(touched_slot_index) {
-                        if let Some(touched_slot) = touched_slot.take_original_ref() {
-                            let slot_value = slot.value.clone();
-                            if touched_slot.value != slot_value {
-                                conflicts.add_slot(address.clone(), touched_slot_index.clone(), slot_value, touched_slot.value.clone());
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        conflicts.build()
-    }
-}
-
-#[derive(Debug, Default)]
-struct InMemoryPermanentStorageState {
-    accounts: HashMap<Address, InMemoryAccountPermanent>,
-    transactions: HashMap<Hash, TransactionMined>,
-    blocks_by_number: IndexMap<BlockNumber, Arc<Block>>,
-    blocks_by_hash: IndexMap<Hash, Arc<Block>>,
-    logs: Vec<LogMined>,
-}
-
-impl StorageState<InMemoryAccountPermanent> for InMemoryPermanentStorageState {
-    fn get_accounts(&self) -> &HashMap<Address, InMemoryAccountPermanent> {
-        &self.accounts
-    }
-
-    fn get_accounts_mut(&mut self) -> &mut HashMap<Address, InMemoryAccountPermanent> {
-        &mut self.accounts
-    }
-}
 
 impl Default for InMemoryStoragePermanent {
     fn default() -> Self {
@@ -371,108 +293,6 @@ impl PermanentStorage for InMemoryStoragePermanent {
             account.reset_at(block_number);
         }
 
-        Ok(())
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct InMemoryStorageTemporary {
-    state: RwLock<InMemoryTemporaryStorageState>,
-}
-
-#[derive(Debug, Default)]
-struct InMemoryTemporaryStorageState {
-    accounts: HashMap<Address, InMemoryAccountTemporary>,
-}
-
-impl StorageState<InMemoryAccountTemporary> for InMemoryTemporaryStorageState {
-    fn get_accounts(&self) -> &HashMap<Address, InMemoryAccountTemporary> {
-        &self.accounts
-    }
-
-    fn get_accounts_mut(&mut self) -> &mut HashMap<Address, InMemoryAccountTemporary> {
-        &mut self.accounts
-    }
-}
-
-impl InMemoryStorageTemporary {
-    /// Locks inner state for reading.
-    async fn lock_read(&self) -> RwLockReadGuard<'_, InMemoryTemporaryStorageState> {
-        self.state.read().await
-    }
-
-    /// Locks inner state for writing.
-    async fn lock_write(&self) -> RwLockWriteGuard<'_, InMemoryTemporaryStorageState> {
-        self.state.write().await
-    }
-}
-
-#[async_trait]
-impl TemporaryStorage for InMemoryStorageTemporary {
-    // -------------------------------------------------------------------------
-    // State operations
-    // ------------------------------------------------------------------------
-
-    async fn check_conflicts(&self, execution: &Execution) -> anyhow::Result<Option<ExecutionConflicts>> {
-        let state_lock = self.lock_read().await;
-        Ok(state_lock.check_conflicts(execution))
-    }
-
-    async fn maybe_read_account(&self, address: &Address, _point_in_time: &StoragePointInTime) -> anyhow::Result<Option<Account>> {
-        tracing::debug!(%address, "reading account");
-
-        let state = self.lock_read().await;
-
-        match state.accounts.get(address) {
-            Some(account) => {
-                let account = Account {
-                    address: address.clone(),
-                    balance: account.balance.clone(),
-                    nonce: account.nonce.clone(),
-                    bytecode: account.bytecode.clone(),
-                };
-                tracing::trace!(%address, ?account, "account found");
-                Ok(Some(account))
-            }
-
-            None => {
-                tracing::trace!(%address, "account not found");
-                Ok(None)
-            }
-        }
-    }
-
-    async fn maybe_read_slot(&self, address: &Address, slot_index: &SlotIndex, point_in_time: &StoragePointInTime) -> anyhow::Result<Option<Slot>> {
-        tracing::debug!(%address, %slot_index, ?point_in_time, "reading slot");
-
-        let state = self.lock_read().await;
-        let Some(account) = state.accounts.get(address) else {
-            tracing::trace!(%address, "account not found");
-            return Ok(Default::default());
-        };
-
-        match account.slots.get(slot_index) {
-            Some(slot) => {
-                tracing::trace!(%address, %slot_index, ?point_in_time, %slot, "slot found");
-                Ok(Some(slot.clone()))
-            }
-
-            None => {
-                tracing::trace!(%address, %slot_index, ?point_in_time, "slot not found");
-                Ok(None)
-            }
-        }
-    }
-
-    async fn save_account_changes(&self, number: BlockNumber, execution: Execution) -> anyhow::Result<()> {
-        let mut state_lock = self.lock_write().await;
-        state_lock.save_account_changes(number, execution);
-        Ok(())
-    }
-
-    async fn reset(&self) -> anyhow::Result<()> {
-        let mut state = self.lock_write().await;
-        state.accounts.clear();
         Ok(())
     }
 }
